Rename global-atoms to global-state
