+++ src/components/command-menu.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from "react-router-dom"
 import { useEvent } from "react-use"
 import { Card } from "../components/card"
 import { PanelsContext } from "../components/panels"
-import { tagSearcherAtom } from "../global-atoms"
+import { tagSearcherAtom } from "../global-state"
 import { templateSchema } from "../types"
 import { formatDate, formatDateDistance } from "../utils/date"
 import { useUpsertNote } from "../utils/github-sync"
