+++ src/components/github-auth.tsx
@@ -2,7 +2,7 @@ import { useAtom, useAtomValue, useSetAtom } from "jotai"
 import React from "react"
 import { useNavigate } from "react-router-dom"
 import urlcat from "urlcat"
-import { githubRepoAtom, githubUserAtom } from "../global-atoms"
+import { githubRepoAtom, githubUserAtom } from "../global-state"
 import { cx } from "../utils/cx"
 import { shaAtom } from "../utils/github-sync"
 import { Button } from "./button"
