+++ src/components/new-note-dialog.tsx
@@ -7,7 +7,7 @@ import React from "react"
 import { DraggableCore } from "react-draggable"
 import { useLocation, useNavigate } from "react-router-dom"
 import { useEvent, useMedia } from "react-use"
-import { githubRepoAtom, githubUserAtom } from "../global-atoms"
+import { githubRepoAtom, githubUserAtom } from "../global-state"
 import { openNewWindow } from "../utils/open-new-window"
 import { useIsFullscreen } from "../utils/use-is-fullscreen"
 import { IconButton } from "./icon-button"
