+++ src/components/note-card-form.tsx
@@ -3,7 +3,7 @@ import { ReactCodeMirrorRef } from "@uiw/react-codemirror"
 import clsx from "clsx"
 import { useAtomValue } from "jotai"
 import React from "react"
-import { githubRepoAtom } from "../global-atoms"
+import { githubRepoAtom } from "../global-state"
 import { NoteId } from "../types"
 import { useUpsertNote } from "../utils/github-sync"
 import { useAttachFile } from "../utils/use-attach-file"
