+++ src/components/note-editor.tsx
@@ -13,7 +13,7 @@ import { parseDate } from "chrono-node"
 import { useAtomCallback } from "jotai/utils"
 import * as emoji from "node-emoji"
 import React from "react"
-import { tagsAtom, templatesAtom } from "../global-atoms"
+import { tagsAtom, templatesAtom } from "../global-state"
 import { formatDate, formatDateDistance } from "../utils/date"
 import { useUpsertNote } from "../utils/github-sync"
 import { parseFrontmatter } from "../utils/parse-frontmatter"
