+++ src/components/root-layout.tsx
@@ -2,7 +2,7 @@ import { useAtomValue } from "jotai"
 import React from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 import { useEvent, useNetworkState } from "react-use"
-import { githubRepoAtom, githubUserAtom } from "../global-atoms"
+import { githubRepoAtom, githubUserAtom } from "../global-state"
 import { useFetchNotes } from "../utils/github-sync"
 import { getPrevPathParams, savePathParams } from "../utils/prev-path-params"
 import { Card } from "./card"
