+++ src/pages/note.tsx
@@ -13,7 +13,7 @@ import { FullscreenNoteForm } from "../components/fullscreen-note-form"
 import { CopyIcon16, EditIcon16, ExternalLinkIcon16, NoteIcon16 } from "../components/icons"
 import { Markdown } from "../components/markdown"
 import { Panels } from "../components/panels"
-import { githubRepoAtom } from "../global-atoms"
+import { githubRepoAtom } from "../global-state"
 import { NotePanel } from "../panels/note"
 import { useUpsertNote } from "../utils/github-sync"
 import { useIsFullscreen } from "../utils/use-is-fullscreen"
