+++ src/pages/settings.tsx
@@ -4,7 +4,7 @@ import { SignedInUser } from "../components/github-auth"
 import { SettingsIcon24 } from "../components/icons"
 import { Panel } from "../components/panel"
 import { RepoForm } from "../components/repo-form"
-import { githubRepoAtom, githubUserAtom } from "../global-atoms"
+import { githubRepoAtom, githubUserAtom } from "../global-state"
 import React from "react"
 import { Button } from "../components/button"
 import { Card } from "../components/card"
