+++ src/panels/tags.tsx
@@ -8,7 +8,7 @@ import { Panel } from "../components/panel"
 import { PanelProps } from "../components/panels"
 import { PillButton } from "../components/pill-button"
 import { SearchInput } from "../components/search-input"
-import { sortedTagEntriesAtom, tagSearcherAtom } from "../global-atoms"
+import { sortedTagEntriesAtom, tagSearcherAtom } from "../global-state"
 import { cx } from "../utils/cx"
 import { pluralize } from "../utils/pluralize"
 import { useSearchParam } from "../utils/use-search-param"
